+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/RequestParametersFromSerializer.kt
@@ -1,6 +1,16 @@
 package at.asitplus.openid
 
 import at.asitplus.dcapi.request.DCAPIRequest
+import at.asitplus.dcapi.request.Oid4vpDCAPIRequest
+import at.asitplus.openid.RequestParametersFrom.SerialNames.DC_API_REQUEST
+import at.asitplus.openid.RequestParametersFrom.SerialNames.JSON_STRING
+import at.asitplus.openid.RequestParametersFrom.SerialNames.JWS_SIGNED
+import at.asitplus.openid.RequestParametersFrom.SerialNames.PARAMETERS
+import at.asitplus.openid.RequestParametersFrom.SerialNames.TYPE_JSON
+import at.asitplus.openid.RequestParametersFrom.SerialNames.TYPE_JWS_SIGNED
+import at.asitplus.openid.RequestParametersFrom.SerialNames.TYPE_URI
+import at.asitplus.openid.RequestParametersFrom.SerialNames.URL
+import at.asitplus.openid.RequestParametersFrom.SerialNames.VERIFIED
 import at.asitplus.signum.indispensable.josef.JwsSigned
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.SerialDescriptor
@@ -15,17 +25,6 @@ import kotlinx.serialization.json.decodeFromJsonElement
 import kotlinx.serialization.json.encodeToJsonElement
 import kotlinx.serialization.json.jsonObject
 
-private const val JwsSignedSerialName = "JwsSigned"
-private const val JsonSerialName = "Json"
-private const val UriSerialName = "Uri"
-
-private const val JwsSignedElementName = "jwsSigned"
-private const val JsonElementName = "jsonString"
-private const val UriElementName = "url"
-private const val ParametersElementName = "parameters"
-private const val DcApiRequestElementName = "dcApiRequest"
-private const val VerifiedElementName = "verified"
-
 /**
  * In order to de-/serialize generic types we need a kind of factory approach
  * Because we deal with a sealed class we can use an intermediary jsonSerializer,
@@ -39,22 +38,22 @@ private const val VerifiedElementName = "verified"
 class RequestParametersFromSerializer<T : RequestParameters>(
     private val parameterSerializer: KSerializer<T>,
 ) : KSerializer<RequestParametersFrom<T>> {
-    val dcApiRequestSerializer = DCAPIRequest.serializer()
+    val dcApiRequestSerializer = Oid4vpDCAPIRequest.serializer()
     override val descriptor: SerialDescriptor = buildClassSerialDescriptor("RequestParametersFromClass") {
-        element(JsonSerialName, buildClassSerialDescriptor(JsonSerialName) {
-            element<String>(JsonElementName)
-            element(ParametersElementName, parameterSerializer.descriptor)
-            element(DcApiRequestElementName, dcApiRequestSerializer.descriptor)
+        element(TYPE_JSON, buildClassSerialDescriptor(TYPE_JSON) {
+            element<String>(JSON_STRING)
+            element(PARAMETERS, parameterSerializer.descriptor)
+            element(DC_API_REQUEST, dcApiRequestSerializer.descriptor)
         })
-        element(JwsSignedSerialName, buildClassSerialDescriptor(JwsSignedSerialName) {
-            element(JwsSignedElementName, JwsSignedSerializer.descriptor)
-            element(ParametersElementName, parameterSerializer.descriptor)
-            element<Boolean>(VerifiedElementName)
-            element(DcApiRequestElementName, dcApiRequestSerializer.descriptor)
+        element(TYPE_JWS_SIGNED, buildClassSerialDescriptor(TYPE_JWS_SIGNED) {
+            element(JWS_SIGNED, JwsSignedSerializer.descriptor)
+            element(PARAMETERS, parameterSerializer.descriptor)
+            element<Boolean>(VERIFIED)
+            element(DC_API_REQUEST, dcApiRequestSerializer.descriptor)
         })
-        element(UriSerialName, buildClassSerialDescriptor(UriSerialName) {
-            element(UriElementName, UrlSerializer.descriptor)
-            element(ParametersElementName, parameterSerializer.descriptor)
+        element(TYPE_URI, buildClassSerialDescriptor(TYPE_URI) {
+            element(URL, UrlSerializer.descriptor)
+            element(PARAMETERS, parameterSerializer.descriptor)
         })
     }
 
@@ -63,32 +62,49 @@ class RequestParametersFromSerializer<T : RequestParameters>(
 
         val element = decoder.decodeJsonElement()
         return when {
-            JsonElementName in element.jsonObject -> RequestParametersFrom.Json(
-                decoder.json.decodeFromJsonElement<String>(element.jsonObject[JsonElementName]!!),
-                decoder.json.decodeFromJsonElement(parameterSerializer, element.jsonObject[ParametersElementName]!!),
-                element.jsonObject[DcApiRequestElementName]?.let { decoder.json.decodeFromJsonElement(dcApiRequestSerializer, it) }
+            JWS_SIGNED in element.jsonObject && DC_API_REQUEST in element.jsonObject -> run {
+                val dcApiRequest =
+                    decoder.json.decodeFromJsonElement(dcApiRequestSerializer, element.jsonObject[DC_API_REQUEST]!!)
+                val parameters =
+                    decoder.json.decodeFromJsonElement(parameterSerializer, element.jsonObject[PARAMETERS]!!)
+                val jwsString = decoder.json.decodeFromJsonElement<String>(element.jsonObject[JWS_SIGNED]!!)
+                val jwsGeneric = JwsSigned.deserialize(jwsString).getOrThrow()
+                RequestParametersFrom.DcApiSigned(
+                    dcApiRequest,
+                    parameters,
+                    JwsSigned<T>(jwsGeneric.header, parameters, jwsGeneric.signature, jwsGeneric.plainSignatureInput),
+                )
+            }
+
+            JSON_STRING in element.jsonObject && DC_API_REQUEST in element.jsonObject -> RequestParametersFrom.DcApiUnsigned(
+                decoder.json.decodeFromJsonElement(dcApiRequestSerializer, element.jsonObject[DC_API_REQUEST]!!),
+                decoder.json.decodeFromJsonElement(parameterSerializer, element.jsonObject[PARAMETERS]!!),
+                decoder.json.decodeFromJsonElement<String>(element.jsonObject[JSON_STRING]!!),
             )
 
-            JwsSignedElementName in element.jsonObject -> run {
-                val parameters =
-                    decoder.json.decodeFromJsonElement(parameterSerializer, element.jsonObject[ParametersElementName]!!)
-                val jwsString = decoder.json.decodeFromJsonElement<String>(element.jsonObject[JwsSignedElementName]!!)
-                val verified = element.jsonObject[VerifiedElementName]?.let { decoder.json.decodeFromJsonElement<Boolean>(it) } ?: false
-                val jwsGeneric = JwsSigned.deserialize(jwsString).getOrThrow()
-                val dcApiRequest = element.jsonObject[DcApiRequestElementName]?.let { decoder.json.decodeFromJsonElement(dcApiRequestSerializer, it) }
+            JSON_STRING in element.jsonObject && DC_API_REQUEST !in element.jsonObject -> RequestParametersFrom.Json(
+                decoder.json.decodeFromJsonElement<String>(element.jsonObject[JSON_STRING]!!),
+                decoder.json.decodeFromJsonElement(parameterSerializer, element.jsonObject[PARAMETERS]!!),
+            )
 
+            JWS_SIGNED in element.jsonObject && DC_API_REQUEST !in element.jsonObject -> run {
+                val parameters =
+                    decoder.json.decodeFromJsonElement(parameterSerializer, element.jsonObject[PARAMETERS]!!)
+                val jwsString = decoder.json.decodeFromJsonElement<String>(element.jsonObject[JWS_SIGNED]!!)
+                val jwsGeneric = JwsSigned.deserialize(jwsString).getOrThrow()
+                val verified = element.jsonObject[VERIFIED]?.let { decoder.json.decodeFromJsonElement<Boolean>(it) }
+                    ?: false
 
                 RequestParametersFrom.JwsSigned(
                     JwsSigned<T>(jwsGeneric.header, parameters, jwsGeneric.signature, jwsGeneric.plainSignatureInput),
                     parameters,
                     verified,
-                    dcApiRequest
                 )
             }
 
-            UriElementName in element.jsonObject -> RequestParametersFrom.Uri(
-                decoder.json.decodeFromJsonElement(UrlSerializer, element.jsonObject[UriElementName]!!),
-                decoder.json.decodeFromJsonElement(parameterSerializer, element.jsonObject[ParametersElementName]!!)
+            URL in element.jsonObject -> RequestParametersFrom.Uri(
+                decoder.json.decodeFromJsonElement(UrlSerializer, element.jsonObject[URL]!!),
+                decoder.json.decodeFromJsonElement(parameterSerializer, element.jsonObject[PARAMETERS]!!)
             )
 
             else -> throw NotImplementedError("Unknown RequestParametersFrom subclass. Input: $element")
@@ -99,21 +115,31 @@ class RequestParametersFromSerializer<T : RequestParameters>(
         require(encoder is JsonEncoder) // this class can be decoded only by Json
         val element = when (value) {
             is RequestParametersFrom.Json -> buildJsonObject {
-                put(JsonElementName, encoder.json.encodeToJsonElement(value.jsonString))
-                put(ParametersElementName, encoder.json.encodeToJsonElement(parameterSerializer, value.parameters))
-                value.dcApiRequest?.let { put(DcApiRequestElementName, encoder.json.encodeToJsonElement(DCAPIRequest.serializer(), it)) }
+                put(JSON_STRING, encoder.json.encodeToJsonElement(value.jsonString))
+                put(PARAMETERS, encoder.json.encodeToJsonElement(parameterSerializer, value.parameters))
             }
 
             is RequestParametersFrom.JwsSigned -> buildJsonObject {
-                put(JwsSignedElementName, encoder.json.encodeToJsonElement(value.jwsSigned.serialize()))
-                put(ParametersElementName, encoder.json.encodeToJsonElement(parameterSerializer, value.parameters))
-                put(VerifiedElementName, encoder.json.encodeToJsonElement(value.verified))
-                value.dcApiRequest?.let { put(DcApiRequestElementName, encoder.json.encodeToJsonElement(DCAPIRequest.serializer(), it)) }
+                put(JWS_SIGNED, encoder.json.encodeToJsonElement(value.jwsSigned.serialize()))
+                put(PARAMETERS, encoder.json.encodeToJsonElement(parameterSerializer, value.parameters))
+                put(VERIFIED, encoder.json.encodeToJsonElement(value.verified))
             }
 
             is RequestParametersFrom.Uri -> buildJsonObject {
-                put(UriElementName, encoder.json.encodeToJsonElement(UrlSerializer, value.url))
-                put(ParametersElementName, encoder.json.encodeToJsonElement(parameterSerializer, value.parameters))
+                put(URL, encoder.json.encodeToJsonElement(UrlSerializer, value.url))
+                put(PARAMETERS, encoder.json.encodeToJsonElement(parameterSerializer, value.parameters))
+            }
+
+            is RequestParametersFrom.DcApiSigned<*> -> buildJsonObject {
+                put(DC_API_REQUEST, encoder.json.encodeToJsonElement(dcApiRequestSerializer, value.dcApiRequest))
+                put(JWS_SIGNED, encoder.json.encodeToJsonElement(value.jwsSigned.serialize()))
+                put(PARAMETERS, encoder.json.encodeToJsonElement(parameterSerializer, value.parameters))
+            }
+
+            is RequestParametersFrom.DcApiUnsigned<*> -> buildJsonObject {
+                put(DC_API_REQUEST, encoder.json.encodeToJsonElement(dcApiRequestSerializer, value.dcApiRequest))
+                put(JSON_STRING, encoder.json.encodeToJsonElement(value.jsonString))
+                put(PARAMETERS, encoder.json.encodeToJsonElement(parameterSerializer, value.parameters))
             }
         }
         encoder.encodeJsonElement(element)